requirements.txt
pydantic==1.10.1
fastapi==0.82.0
uvicorn==0.18.3
python-dotenv==0.21.0
sqlalchemy==1.4.41
databases==0.6.1
asyncpg==0.26.0